'createBrowserContextAndPageAndLoginPageInstances' moved to upper class 'BaseTest'
